utilizando DTOs e tratando regras de negocios

+++ src/main/java/poke/api/service/PokemonService.java
@@ -1,9 +1,14 @@
 package poke.api.service;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import poke.api.controller.PokemonController;
 import poke.api.model.Pokemon;
+import poke.api.model.dto.PokemonRequestDTO;
+import poke.api.model.dto.converter.PokemonConverter;
 import poke.api.repository.PokemonRepository;
 
 import java.util.List;
@@ -13,14 +18,15 @@ public class PokemonService {
 
     private PokemonRepository pokemonRepository;
 
-    public  PokemonService(PokemonRepository pokemonRepository){
+    public PokemonService(PokemonRepository pokemonRepository) {
         this.pokemonRepository = pokemonRepository;
     }
-    public List<Pokemon> buscarTodos(){
+
+    public List<Pokemon> buscarTodos() {
         return pokemonRepository.findAll();
     }
 
-    public Pokemon buscarPeloNome(String nome){
+    public Pokemon buscarPeloNome(String nome) {
         return pokemonRepository.findByNome(nome);
     }
 
@@ -45,20 +51,39 @@ public class PokemonService {
         return pokemonRepository.findById(id).orElse(null);
     }
 
-    public void adicionar(Pokemon pokemon){
-        pokemonRepository.save(pokemon);
+    public void adicionar(PokemonRequestDTO pokemonRequestDTO) {
+        String nomePokemonAdicionado = pokemonRequestDTO.getNome();
+        Pokemon pokemonExistente = buscarPeloNome(nomePokemonAdicionado);
+        //Verificar se pokemon ja existe no banco de dados
+        if (pokemonExistente != null) {
+            //Retornar HTTP Status code 400 com mensagem explicativa
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Não é possivel adicionar um pokemon repetido. O pokemon "
+                    + nomePokemonAdicionado + " ja existe no banco de dados.");
+        } else {
+            Pokemon pokemonEntity = PokemonConverter.converterParaEntidade(pokemonRequestDTO);
+            pokemonRepository.save(pokemonEntity);
+        }
     }
 
-    public Pokemon removerPorId(Long id){
+    public Pokemon removerPorId(Long id) {
         Pokemon pokemonParaRemover = pokemonRepository.findById(id).get();
         pokemonRepository.delete(pokemonParaRemover);
-       return pokemonParaRemover;
+        return pokemonParaRemover;
     }
 
-    public Pokemon removerPorNome(String nome){
+    public Pokemon removerPorNome(String nome) {
         Pokemon pokemonRemoverPorNome = pokemonRepository.findByNome(nome);
-        pokemonRepository.delete(pokemonRemoverPorNome);
+        if (pokemonRemoverPorNome == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Não é possivel remover um pokemon inexistente. O pokemon " + nome + " não existe no banco de dados.");
+        } else {
+            pokemonRepository.delete(pokemonRemoverPorNome);
+
+        }
         return pokemonRemoverPorNome;
+    }
+
+    public Long contar(){
+        return pokemonRepository.count();
     }
